+++ changelog.md
@@ -1,3 +1,8 @@
+# 1.0.4
+
+- Add `from_config` static method.
+- Add `get_chapter_by_path` method.
+
 # 1.0.3
 
 -   Add `get_chapter_title` method.